requirements.txt
streamlit
pandas
matplotlib
seaborn
langchain
langchain-groq
langchain-experimental
python-dotenv